Fixed leap year bug
Before, if the current year was a leap year, one day was automatically added. However, this is not always correct. I added a condition that ensures the 29 February has passed. If it is has not, then it is not added. Also something that should be noted, if the day is the 29 February, the extra day is not added, as the day section would handle this (29 - 28 = 1).