Ensures that input is possible
Now, a date must actually be possible e.g. 30/02 is not longer allowed. (It takes in account leap years for February.) Other than that, a future date cannot be entered e.g. 19/06/2022 (tomorrow). Before, future dates that were the same year as the current year were allowed. That meant, the previous date was allowed but 19/06/2023 was not, for example.