Fixed bug caused by previous leap year bug fix
The last edit on leap years added the conditon that it be past the leap year in the current year -- but this applied to all years. That means, if it is January this year, then it won't add ANY leap days whatsoever, even ones from past years. The fix here was very simple -- if not from this year, then go ahead even if it not passed January.